requirements.txt
pandas>=1.3.0
numpy>=1.20.0
fyers-apiv3>=3.0.0
openpyxl>=3.0.0
plotly>=5.0.0
streamlit>=1.20.0
setuptools>=65.0.0

pyyaml>=6.0.0

websocket-client>=1.7.0
